Create README.md
The "4-Queens Problem" is a simplified version of the more general "N-Queens Problem", where the goal is to place N  queens on an N times N  chessboard such that no two queens threaten each other. This puzzle is widely used to teach concepts of constraint satisfaction, algorithmic problem-solving, and backtracking.